Moving summary card to styles folder
