+++ upnext-web-app/src/UpNext/components/styles/SummaryCard.css
